feat: add telemetry to runtimes and flow start
